There is no "star 86"
GUIDO refers to star 86 but there was no such thing. The list only went up to 50, including Sun, Earth, and Moon.
It might refer to "dead reckoning", as NOUN 86 would hold a Delta V vector;
Or it might refer to VERB 86 which would reject a mark taken on the backup optics; however, that was in the lunar rendezvous routine.